refactor: Round 3 — character_agent and wrestler_lifecycle decomposed
character_agent.py (891 → 500 lines, -44%):
  - Extract character_prompts.py (495 lines) — all prompt templates,
    system messages, personality dicts, and prompt builder functions
  - DRY up 3 identical response parsing blocks into _parse_tagged_response()
  - Main file now focused on LLM orchestration, not string templates

wrestler_lifecycle_service.py (787 → 581 lines, -26%):
  - Extract politics_service.py (115 lines) — locker room dynamics
  - lifecycle_constants.py and goal_service.py already extracted in
    previous commit

No behavior changes. All Python syntax validated.
