+++ game_service/wrestler_lifecycle_service.py
@@ -3,25 +3,87 @@ Wrestler Lifecycle Service — Groups 1-6
 
 Handles: aging & decline, career goals, backstage politics,
 developmental pipeline, legacy/Hall of Fame, physical identity & conditioning.
+
+Group 2 (goals) lives in goal_service.py.
+Group 3 (politics) lives in politics_service.py.
+Constants live in lifecycle_constants.py.
 """
 
 import logging
 import random
 from datetime import datetime, timedelta
-from typing import List, Optional
+from typing import Optional
 
 from sqlalchemy.orm import Session
 
 from models.game_models import (
     GameWrestlerDB, WrestlerStatsDB, GameFederationDB, ContractDB,
-    ChampionshipDB, ChampionshipHistoryDB, MatchParticipantDB, MatchDB,
-    ShowDB, ShowSegmentDB, WrestlerPushDB, BookingVisionDB,
-    GameNarrativeLogDB, WrestlerHistoryDB,
-    WrestlerGoalDB, MentorshipDB, CareerHighlightDB, HallOfFameDB,
+    ChampionshipHistoryDB, MatchParticipantDB, MatchDB,
+    ShowDB, ShowSegmentDB, WrestlerPushDB,
+    GameNarrativeLogDB,
+    MentorshipDB, CareerHighlightDB, HallOfFameDB,
     GimmickHistoryDB, WrestlerBackstoryDB, LifeEventDB,
+)
+from game_service.lifecycle_constants import (
+    # Group 1: Aging
+    CAREER_PHASE_ROOKIE_MAX_EXP, CAREER_PHASE_RISING_OFFSET,
+    CAREER_PHASE_PRIME_OFFSET, CAREER_PHASE_VETERAN_OFFSET,
+    MAX_DECLINE_PER_YEAR, MIN_STAT_FLOOR, MAX_STAT_CAP,
+    EARLY_DECLINE_STATS, LATE_DECLINE_STATS, LATE_DECLINE_YEARS_THRESHOLD,
+    IMPROVING_STATS, IMPROVING_STAT_MAX_GAIN,
+    RETIREMENT_PRESSURE,
+    RING_RUST_NO_PENALTY_DAYS, RING_RUST_DIVISOR, RING_RUST_MIN_MODIFIER,
+    # Group 4: Developmental
+    MENTOR_MIN_PSYCHOLOGY, MENTOR_BONUS_DIVISOR,
+    DEBUT_MIN_WEEKS, DEBUT_MIN_AVG_RING, DEBUT_MIN_PSYCHOLOGY,
+    DEBUT_RING_STATS,
+    MENTOR_SPECIALTY_MULTIPLIER, MENTOR_PSYCHOLOGY_BONUS,
+    MENTOR_SELF_IMPROVE_CHANCE,
+    # Group 5: Legacy
+    LEGACY_HIGHLIGHT_WEIGHT, LEGACY_REIGN_WEIGHT,
+    LEGACY_RATING_WEIGHT, LEGACY_YEARS_WEIGHT,
+    HIGHLIGHT_STAR_THRESHOLD, HIGHLIGHT_MAX_SIGNIFICANCE,
+    HIGHLIGHT_SIGNIFICANCE_MULTIPLIER,
+    HOF_MIN_LEGACY,
+    NOSTALGIA_MIN_DAYS_ABSENT, NOSTALGIA_MIN_LEGACY,
+    NOSTALGIA_MAX_BONUS, NOSTALGIA_DAYS_PER_UNIT, NOSTALGIA_PER_UNIT,
+    # Group 6: Physical
+    BODY_TYPE_THRESHOLDS, BODY_TYPE_DEFAULT,
+    HEIGHT_RANGE, WEIGHT_OFFSET_RANGE, WEIGHT_MIN, WEIGHT_MAX,
+    HEIGHT_WEIGHT_FACTOR,
+    BODY_MOD_HEAVY_DIFF, BODY_MOD_HEAVY, BODY_MOD_LIGHT, BODY_MOD_NEUTRAL,
+    STIPULATION_SPECIALIST_MAP, STIPULATION_BONUS_DIVISOR,
+    SPECIALIZATION_GROWTH,
+    CONDITIONING_DEFAULT, CONDITIONING_OVERWORK_THRESHOLD,
+    CONDITIONING_OVERWORK_PENALTY, CONDITIONING_REST_GAIN,
+    CONDITIONING_WORK_GAIN, CONDITIONING_MIN,
+)
+
+# Re-export Group 2 (career goals) so existing imports keep working
+from game_service.goal_service import (          # noqa: F401
+    create_wrestler_goals,
+    evaluate_goals,
+    _check_goal_completed,
+)
+
+# Re-export Group 3 (backstage politics) so existing imports keep working
+from game_service.politics_service import (      # noqa: F401
+    update_locker_room_dynamics,
+    apply_politics_to_booking,
 )
 
 logger = logging.getLogger(__name__)
+
+
+# ---------------------------------------------------------------------------
+# Shared helpers
+# ---------------------------------------------------------------------------
+
+def _get_wrestler_stats(db: Session, wrestler_id: str) -> Optional[WrestlerStatsDB]:
+    """Fetch a wrestler's stats row (returns None if missing)."""
+    return db.query(WrestlerStatsDB).filter(
+        WrestlerStatsDB.wrestler_id == wrestler_id
+    ).first()
 
 
 # ---------------------------------------------------------------------------
@@ -34,13 +96,13 @@ def update_career_phase(wrestler: GameWrestlerDB):
     peak = wrestler.peak_age or 28
     exp = wrestler.experience_years or 0
 
-    if exp < 2:
+    if exp < CAREER_PHASE_ROOKIE_MAX_EXP:
         wrestler.career_phase = "rookie"
-    elif age < peak - 2:
+    elif age < peak + CAREER_PHASE_RISING_OFFSET:
         wrestler.career_phase = "rising"
-    elif age < peak + 4:
+    elif age < peak + CAREER_PHASE_PRIME_OFFSET:
         wrestler.career_phase = "prime"
-    elif age < peak + 8:
+    elif age < peak + CAREER_PHASE_VETERAN_OFFSET:
         wrestler.career_phase = "veteran"
     else:
         wrestler.career_phase = "declining"
@@ -61,53 +123,49 @@ def age_wrestlers(db: Session, world_id: str, game_date: str):
         # Physical stat decline past peak
         peak = w.peak_age or 28
         if w.age > peak:
-            stats = db.query(WrestlerStatsDB).filter(
-                WrestlerStatsDB.wrestler_id == w.id
-            ).first()
+            stats = _get_wrestler_stats(db, w.id)
             if not stats:
                 continue
 
             years_past = w.age - peak
-            decline = min(years_past, 5)  # Max 5 pts/year decline
+            decline = min(years_past, MAX_DECLINE_PER_YEAR)
 
-            # Physical stats decay first
-            for attr in ("speed", "aerial", "stamina"):
+            for attr in EARLY_DECLINE_STATS:
                 old = getattr(stats, attr, 50)
-                setattr(stats, attr, max(5, old - random.randint(1, decline)))
+                setattr(stats, attr, max(MIN_STAT_FLOOR, old - random.randint(1, decline)))
 
-            # Power/toughness decay later
-            if years_past > 3:
-                for attr in ("power", "toughness"):
+            if years_past > LATE_DECLINE_YEARS_THRESHOLD:
+                for attr in LATE_DECLINE_STATS:
                     old = getattr(stats, attr, 50)
-                    setattr(stats, attr, max(5, old - random.randint(0, decline - 1)))
+                    setattr(stats, attr, max(MIN_STAT_FLOOR, old - random.randint(0, decline - 1)))
 
-            # Mental stats IMPROVE with age
-            for attr in ("psychology", "selling"):
+            for attr in IMPROVING_STATS:
                 old = getattr(stats, attr, 50)
-                setattr(stats, attr, min(100, old + random.randint(0, 2)))
+                setattr(stats, attr, min(MAX_STAT_CAP, old + random.randint(0, IMPROVING_STAT_MAX_GAIN)))
 
 
 def calculate_retirement_pressure(wrestler: GameWrestlerDB) -> int:
     """Calculate how likely a wrestler is to retire (replaces flat 1% random)."""
+    rp = RETIREMENT_PRESSURE
     pressure = 0
     phase = wrestler.career_phase or "prime"
 
     if phase == "declining":
-        pressure += 10
+        pressure += rp["declining_phase"]
     elif phase == "veteran":
-        pressure += 3
+        pressure += rp["veteran_phase"]
 
-    if (wrestler.morale or 50) < 30:
-        pressure += 15
+    if (wrestler.morale or 50) < rp["low_morale_threshold"]:
+        pressure += rp["low_morale_bonus"]
 
-    if wrestler.is_injured and (wrestler.age or 25) > 35:
-        pressure += 20
+    if wrestler.is_injured and (wrestler.age or 25) > rp["injured_age_threshold"]:
+        pressure += rp["injured_bonus"]
 
-    if (wrestler.popularity or 50) < 20 and (wrestler.age or 25) > 36:
-        pressure += 10
+    if (wrestler.popularity or 50) < rp["low_pop_threshold"] and (wrestler.age or 25) > rp["low_pop_age_threshold"]:
+        pressure += rp["low_pop_bonus"]
 
-    if (wrestler.age or 25) > 42:
-        pressure += 15
+    if (wrestler.age or 25) > rp["old_age_threshold"]:
+        pressure += rp["old_age_bonus"]
 
     return pressure
 
@@ -115,220 +173,9 @@ def calculate_retirement_pressure(wrestler: GameWrestlerDB) -> int:
 def calculate_ring_rust_modifier(wrestler: GameWrestlerDB) -> float:
     """Ring rust modifier for match engine (0.85 - 1.0)."""
     rust = wrestler.ring_rust_days or 0
-    if rust <= 14:
+    if rust <= RING_RUST_NO_PENALTY_DAYS:
         return 1.0
-    return max(0.85, 1.0 - (rust / 500))
-
-
-# ---------------------------------------------------------------------------
-# Group 2: Career Goals
-# ---------------------------------------------------------------------------
-
-def create_wrestler_goals(db: Session, wrestler: GameWrestlerDB, game_date: str):
-    """Create structured goal records from the wrestler's career_goals JSON."""
-    goals = wrestler.career_goals or []
-    for goal_type in goals:
-        existing = db.query(WrestlerGoalDB).filter(
-            WrestlerGoalDB.wrestler_id == wrestler.id,
-            WrestlerGoalDB.goal_type == goal_type,
-            WrestlerGoalDB.status == "active",
-        ).first()
-        if not existing:
-            db.add(WrestlerGoalDB(
-                wrestler_id=wrestler.id,
-                goal_type=goal_type,
-                set_date=game_date,
-            ))
-
-
-def evaluate_goals(db: Session, wrestler: GameWrestlerDB, game_date: str):
-    """Check progress on all active goals. Returns list of completed goal types."""
-    active = db.query(WrestlerGoalDB).filter(
-        WrestlerGoalDB.wrestler_id == wrestler.id,
-        WrestlerGoalDB.status == "active",
-    ).all()
-
-    completed = []
-    for goal in active:
-        if _check_goal_completed(db, wrestler, goal):
-            goal.status = "completed"
-            goal.completed_date = game_date
-            completed.append(goal.goal_type)
-
-            wrestler.morale = min(100, (wrestler.morale or 50) + 10)
-            wrestler.satisfaction = min(100, (wrestler.satisfaction or 50) + 15)
-
-            db.add(WrestlerHistoryDB(
-                wrestler_id=wrestler.id,
-                game_date=game_date,
-                event_type="goal_completed",
-                description=f"Achieved career goal: {goal.goal_type}",
-            ))
-        else:
-            # Frustration grows when stuck
-            goal.frustration = min(100, (goal.frustration or 0) + 1)
-
-            if goal.frustration > 80:
-                wrestler.morale = max(0, (wrestler.morale or 50) - 3)
-                # Frustrated faces drift heel
-                if wrestler.alignment == "face" and goal.frustration > 90:
-                    wrestler.alignment_momentum = (wrestler.alignment_momentum or 0) - 5
-
-    # Glass ceiling detection
-    push = db.query(WrestlerPushDB).filter(
-        WrestlerPushDB.wrestler_id == wrestler.id,
-    ).first()
-    if push and (push.weeks_at_tier or 0) > 26:
-        title_goals = [g for g in active if g.goal_type in (
-            "win_title", "become_champion", "win_first_title",
-            "main_event_ppv", "one_more_title_run",
-        )]
-        for g in title_goals:
-            g.frustration = min(100, (g.frustration or 0) + 3)
-
-    # Satisfaction influences morale
-    sat = wrestler.satisfaction or 50
-    if sat < 30:
-        wrestler.morale = max(0, (wrestler.morale or 50) - 2)
-    elif sat > 70:
-        wrestler.morale = min(100, (wrestler.morale or 50) + 1)
-
-    return completed
-
-
-def _check_goal_completed(db: Session, wrestler: GameWrestlerDB, goal: WrestlerGoalDB) -> bool:
-    """Check if a specific goal has been achieved."""
-    gt = goal.goal_type
-
-    if gt in ("win_title", "become_champion", "win_first_title", "one_more_title_run"):
-        # Check if wrestler currently holds any title
-        champ = db.query(ChampionshipDB).filter(
-            ChampionshipDB.current_holder_id == wrestler.id,
-        ).first()
-        return champ is not None
-
-    if gt == "main_event_ppv":
-        # Check if wrestler main evented a PPV show
-        ppv_main = db.query(MatchParticipantDB).join(MatchDB).join(ShowSegmentDB).join(ShowDB).filter(
-            MatchParticipantDB.wrestler_id == wrestler.id,
-            ShowDB.show_type == "ppv",
-            MatchDB.is_completed == True,
-        ).first()
-        return ppv_main is not None
-
-    if gt in ("have_5_star_match", "5_star_match"):
-        best = db.query(MatchParticipantDB).join(MatchDB).filter(
-            MatchParticipantDB.wrestler_id == wrestler.id,
-            MatchDB.match_rating >= 4.8,
-        ).first()
-        return best is not None
-
-    if gt == "defeat_rival":
-        if goal.target_entity_id:
-            win = db.query(MatchParticipantDB).join(MatchDB).filter(
-                MatchParticipantDB.wrestler_id == wrestler.id,
-                MatchParticipantDB.is_winner == True,
-                MatchDB.winner_id == wrestler.id,
-            ).first()
-            return win is not None
-        return False
-
-    if gt == "prove_myself":
-        return (wrestler.popularity or 0) >= 50
-
-    if gt in ("build_legacy", "cement_legacy"):
-        return (wrestler.legacy_score or 0) >= 50
-
-    if gt == "become_top_draw":
-        return (wrestler.draw_rating or 0) >= 80
-
-    if gt in ("earn_respect", "prove_doubters_wrong"):
-        return (wrestler.popularity or 0) >= 60 and (wrestler.morale or 0) >= 60
-
-    if gt == "make_it_to_main_event":
-        push = db.query(WrestlerPushDB).filter(
-            WrestlerPushDB.wrestler_id == wrestler.id,
-            WrestlerPushDB.push_tier == "main_event",
-        ).first()
-        return push is not None
-
-    if gt == "headline_biggest_show":
-        return (wrestler.popularity or 0) >= 75
-
-    return False
-
-
-# ---------------------------------------------------------------------------
-# Group 3: Backstage Politics & Locker Room Power
-# ---------------------------------------------------------------------------
-
-def update_locker_room_dynamics(db: Session, federation: GameFederationDB, game_date: str):
-    """Weekly locker room standing/influence calculation and morale contagion."""
-    contracts = db.query(ContractDB).filter(
-        ContractDB.federation_id == federation.id,
-        ContractDB.status == "active",
-    ).all()
-    wrestler_ids = [c.wrestler_id for c in contracts]
-    if not wrestler_ids:
-        return
-
-    wrestlers = db.query(GameWrestlerDB).filter(
-        GameWrestlerDB.id.in_(wrestler_ids),
-        GameWrestlerDB.is_active == True,
-    ).all()
-
-    leaders = []
-    toxic_count = 0
-
-    for w in wrestlers:
-        stats = db.query(WrestlerStatsDB).filter(
-            WrestlerStatsDB.wrestler_id == w.id
-        ).first()
-        if not stats:
-            continue
-
-        politics = stats.backstage_politics or 50
-        tenure = min((w.experience_years or 0) * 2, 20)
-        pop = (w.popularity or 50) // 5
-        w.creative_influence = min(100, politics + tenure + pop)
-
-        work = stats.work_ethic or 50
-        if politics > 80 and (w.popularity or 0) > 70:
-            w.locker_room_standing = "leader"
-            leaders.append(w)
-        elif politics > 60 and work > 60:
-            w.locker_room_standing = "respected"
-        elif politics < 30 and work < 40:
-            w.locker_room_standing = "disliked"
-        elif politics > 70 and work < 30:
-            w.locker_room_standing = "toxic"
-            toxic_count += 1
-        else:
-            w.locker_room_standing = "neutral"
-
-    # Morale contagion
-    if leaders:
-        avg_leader_morale = sum(l.morale or 50 for l in leaders) / len(leaders)
-        shift = int((avg_leader_morale - 50) / 20)  # -2 to +2
-        for w in wrestlers:
-            if w not in leaders and shift != 0:
-                w.morale = max(0, min(100, (w.morale or 50) + shift))
-
-    if toxic_count > 0:
-        for w in wrestlers:
-            if (w.locker_room_standing or "neutral") != "toxic":
-                w.morale = max(0, (w.morale or 50) - toxic_count)
-
-
-def apply_politics_to_booking(db: Session, wrestler: GameWrestlerDB,
-                              planned_finish: str) -> str:
-    """Creative control: high-influence veterans may refuse clean losses.
-    Returns potentially modified finish type."""
-    influence = wrestler.creative_influence or 0
-    if influence > 75 and planned_finish in ("pinfall", "submission"):
-        if random.random() < influence / 200:
-            return random.choice(["count_out", "disqualification"])
-    return planned_finish
+    return max(RING_RUST_MIN_MODIFIER, 1.0 - (rust / RING_RUST_DIVISOR))
 
 
 # ---------------------------------------------------------------------------
@@ -339,13 +186,11 @@ def assign_mentor(db: Session, federation: GameFederationDB,
                   protege: GameWrestlerDB, mentor: GameWrestlerDB,
                   game_date: str) -> Optional[MentorshipDB]:
     """Assign a veteran mentor to a young wrestler."""
-    mentor_stats = db.query(WrestlerStatsDB).filter(
-        WrestlerStatsDB.wrestler_id == mentor.id
-    ).first()
-    if not mentor_stats or (mentor_stats.psychology or 0) < 50:
+    mentor_stats = _get_wrestler_stats(db, mentor.id)
+    if not mentor_stats or (mentor_stats.psychology or 0) < MENTOR_MIN_PSYCHOLOGY:
         return None
 
-    bonus = ((mentor_stats.psychology or 50) + (mentor_stats.work_ethic or 50)) / 200
+    bonus = ((mentor_stats.psychology or 50) + (mentor_stats.work_ethic or 50)) / MENTOR_BONUS_DIVISOR
     m = MentorshipDB(
         world_id=federation.world_id,
         mentor_id=mentor.id,
@@ -369,7 +214,6 @@ def auto_assign_mentors(db: Session, federation: GameFederationDB, game_date: st
     if not wrestler_ids:
         return
 
-    # Find rookies without mentors
     rookies = db.query(GameWrestlerDB).filter(
         GameWrestlerDB.id.in_(wrestler_ids),
         GameWrestlerDB.is_active == True,
@@ -390,7 +234,6 @@ def auto_assign_mentors(db: Session, federation: GameFederationDB, game_date: st
         ).all()
     )
 
-    # Find available veterans
     veterans = db.query(GameWrestlerDB).filter(
         GameWrestlerDB.id.in_(wrestler_ids),
         GameWrestlerDB.is_active == True,
@@ -417,21 +260,15 @@ def check_debut_readiness(db: Session, wrestler: GameWrestlerDB) -> bool:
         WrestlerPushDB.wrestler_id == wrestler.id,
         WrestlerPushDB.push_tier == "developmental",
     ).first()
-    if not push or (push.weeks_at_tier or 0) < 8:
+    if not push or (push.weeks_at_tier or 0) < DEBUT_MIN_WEEKS:
         return False
 
-    stats = db.query(WrestlerStatsDB).filter(
-        WrestlerStatsDB.wrestler_id == wrestler.id
-    ).first()
+    stats = _get_wrestler_stats(db, wrestler.id)
     if not stats:
         return False
 
-    avg_ring = (
-        (stats.power or 50) + (stats.technical or 50) +
-        (stats.aerial or 50) + (stats.brawling or 50)
-    ) / 4
-
-    return avg_ring > 45 and (stats.psychology or 0) > 35
+    avg_ring = sum(getattr(stats, attr, 50) or 50 for attr in DEBUT_RING_STATS) / len(DEBUT_RING_STATS)
+    return avg_ring > DEBUT_MIN_AVG_RING and (stats.psychology or 0) > DEBUT_MIN_PSYCHOLOGY
 
 
 def training_with_mentor(db: Session, wrestler_id: str, stat_name: str) -> int:
@@ -445,19 +282,15 @@ def training_with_mentor(db: Session, wrestler_id: str, stat_name: str) -> int:
         return 0
 
     bonus = 0
-    # Mentor's specialty gives extra bonus
     if stat_name == mentorship.skill_focus:
-        bonus = max(1, int(mentorship.mentor_bonus * 2))
+        bonus = max(1, int(mentorship.mentor_bonus * MENTOR_SPECIALTY_MULTIPLIER))
     elif stat_name == "psychology":
-        bonus = 1  # Proteges always learn psychology faster
+        bonus = MENTOR_PSYCHOLOGY_BONUS
 
-    # Small chance mentor's psychology improves too (teaching deepens understanding)
-    if random.random() < 0.1:
-        mentor_stats = db.query(WrestlerStatsDB).filter(
-            WrestlerStatsDB.wrestler_id == mentorship.mentor_id
-        ).first()
-        if mentor_stats and (mentor_stats.psychology or 0) < 100:
-            mentor_stats.psychology = min(100, (mentor_stats.psychology or 50) + 1)
+    if random.random() < MENTOR_SELF_IMPROVE_CHANCE:
+        mentor_stats = _get_wrestler_stats(db, mentorship.mentor_id)
+        if mentor_stats and (mentor_stats.psychology or 0) < MAX_STAT_CAP:
+            mentor_stats.psychology = min(MAX_STAT_CAP, (mentor_stats.psychology or 50) + 1)
 
     return bonus
 
@@ -485,20 +318,20 @@ def check_match_highlights(db: Session, match: MatchDB, wrestler_id: str,
     """Check if a match produced career highlights."""
     rating = match.match_rating or 0
 
-    if rating >= 4.5:
+    if rating >= HIGHLIGHT_STAR_THRESHOLD:
         record_career_highlight(
             db, wrestler_id, "5_star_classic",
             f"A {rating}-star classic",
-            game_date, significance=min(10, int(rating * 2)),
+            game_date,
+            significance=min(HIGHLIGHT_MAX_SIGNIFICANCE, int(rating * HIGHLIGHT_SIGNIFICANCE_MULTIPLIER)),
             match_id=match.id,
         )
 
-    # First title win
     if match.is_title_match and match.winner_id == wrestler_id:
         prev_reigns = db.query(ChampionshipHistoryDB).filter(
             ChampionshipHistoryDB.wrestler_id == wrestler_id,
         ).count()
-        if prev_reigns <= 1:  # This is the first reign
+        if prev_reigns <= 1:
             record_career_highlight(
                 db, wrestler_id, "first_title_win",
                 "Won their first championship",
@@ -521,25 +354,26 @@ def compute_legacy_score(db: Session, wrestler_id: str) -> int:
     ).first()
     years = wrestler.experience_years or 0 if wrestler else 0
 
-    # Average match rating
     participations = db.query(MatchParticipantDB).join(MatchDB).filter(
         MatchParticipantDB.wrestler_id == wrestler_id,
         MatchDB.is_completed == True,
         MatchDB.match_rating != None,
     ).all()
-    if participations:
-        avg_rating = sum(
-            p.performance_rating or 3.0 for p in participations
-        ) / len(participations)
-    else:
-        avg_rating = 0
+    avg_rating = (
+        sum(p.performance_rating or 3.0 for p in participations) / len(participations)
+        if participations else 0
+    )
 
-    return int((highlights * 5) + (reigns * 10) + (avg_rating * 8) + (years * 2))
+    return int(
+        (highlights * LEGACY_HIGHLIGHT_WEIGHT)
+        + (reigns * LEGACY_REIGN_WEIGHT)
+        + (avg_rating * LEGACY_RATING_WEIGHT)
+        + (years * LEGACY_YEARS_WEIGHT)
+    )
 
 
 def hall_of_fame_ceremony(db: Session, world_id: str, game_date: str):
     """Annual Hall of Fame induction — run on April 1 each game year."""
-    # Find eligible: retired, not inducted, legacy > 50
     inducted_ids = set(
         h.wrestler_id for h in db.query(HallOfFameDB).filter(
             HallOfFameDB.world_id == world_id,
@@ -559,7 +393,7 @@ def hall_of_fame_ceremony(db: Session, world_id: str, game_date: str):
             continue
         score = compute_legacy_score(db, w.id)
         w.legacy_score = score
-        if score > best_score and score > 50:
+        if score > best_score and score > HOF_MIN_LEGACY:
             best = w
             best_score = score
 
@@ -597,11 +431,11 @@ def apply_nostalgia_pop(wrestler: GameWrestlerDB, game_date: str,
     except (ValueError, TypeError):
         return 0
 
-    if days_absent < 90 or (wrestler.legacy_score or 0) < 30:
+    if days_absent < NOSTALGIA_MIN_DAYS_ABSENT or (wrestler.legacy_score or 0) < NOSTALGIA_MIN_LEGACY:
         return 0
 
-    pop_bonus = min(30, (days_absent // 30) * 5)
-    wrestler.popularity = min(100, (wrestler.popularity or 50) + pop_bonus)
+    pop_bonus = min(NOSTALGIA_MAX_BONUS, (days_absent // NOSTALGIA_DAYS_PER_UNIT) * NOSTALGIA_PER_UNIT)
+    wrestler.popularity = min(MAX_STAT_CAP, (wrestler.popularity or 50) + pop_bonus)
     return pop_bonus
 
 
@@ -611,23 +445,17 @@ def apply_nostalgia_pop(wrestler: GameWrestlerDB, game_date: str,
 
 def derive_body_type(height_cm: int, weight_kg: int) -> str:
     """Derive body_type from height and weight."""
-    bmi_like = weight_kg / (height_cm / 100) ** 2
-    if weight_kg < 85:
-        return "cruiserweight"
-    elif weight_kg < 110:
-        return "average"
-    elif weight_kg < 140:
-        return "big_man"
-    else:
-        return "super_heavyweight"
+    for threshold, btype in BODY_TYPE_THRESHOLDS:
+        if weight_kg < threshold:
+            return btype
+    return BODY_TYPE_DEFAULT
 
 
 def generate_physical_attributes() -> dict:
     """Generate height_cm, weight_kg, body_type for a wrestler."""
-    height = random.randint(165, 205)
-    # Weight correlates loosely with height
-    base_weight = int(height * 0.5 + random.randint(-10, 20))
-    weight = max(70, min(160, base_weight))
+    height = random.randint(*HEIGHT_RANGE)
+    base_weight = int(height * HEIGHT_WEIGHT_FACTOR + random.randint(*WEIGHT_OFFSET_RANGE))
+    weight = max(WEIGHT_MIN, min(WEIGHT_MAX, base_weight))
     return {
         "height_cm": height,
         "weight_kg": weight,
@@ -638,47 +466,30 @@ def generate_physical_attributes() -> dict:
 def calculate_body_modifier(attacker_weight: int, defender_weight: int) -> dict:
     """Calculate stat modifiers based on weight difference."""
     diff = (attacker_weight or 100) - (defender_weight or 100)
-    mods = {"power": 1.0, "speed": 1.0, "aerial": 1.0}
-
-    if diff > 30:  # Much heavier
-        mods["power"] = 1.15
-        mods["aerial"] = 0.85
-    elif diff < -30:  # Much lighter
-        mods["speed"] = 1.15
-        mods["power"] = 0.85
-        mods["aerial"] = 1.10
-
-    return mods
+    if diff > BODY_MOD_HEAVY_DIFF:
+        return dict(BODY_MOD_HEAVY)
+    elif diff < -BODY_MOD_HEAVY_DIFF:
+        return dict(BODY_MOD_LIGHT)
+    return dict(BODY_MOD_NEUTRAL)
 
 
 def calculate_stipulation_bonus(stats: WrestlerStatsDB, stipulation: str) -> float:
     """Stipulation specialist bonus multiplier (1.0 - 1.5)."""
     if not stipulation:
         return 1.0
-
-    mapping = {
-        "cage": stats.cage_specialist or 0,
-        "hell_in_a_cell": stats.cage_specialist or 0,
-        "ladder": stats.ladder_specialist or 0,
-        "tables": stats.hardcore_specialist or 0,
-        "no_dq": stats.hardcore_specialist or 0,
-        "falls_count_anywhere": stats.hardcore_specialist or 0,
-        "extreme_rules": stats.hardcore_specialist or 0,
-        "street_fight": stats.hardcore_specialist or 0,
-    }
-    spec = mapping.get(stipulation.lower().replace(" ", "_"), 0)
-    return 1.0 + (spec / 200)
+    attr = STIPULATION_SPECIALIST_MAP.get(stipulation.lower().replace(" ", "_"))
+    if not attr:
+        return 1.0
+    spec = getattr(stats, attr, 0) or 0
+    return 1.0 + (spec / STIPULATION_BONUS_DIVISOR)
 
 
 def update_conditioning(db: Session, wrestler: GameWrestlerDB, game_date: str):
     """Update conditioning based on recent workload."""
-    stats = db.query(WrestlerStatsDB).filter(
-        WrestlerStatsDB.wrestler_id == wrestler.id
-    ).first()
+    stats = _get_wrestler_stats(db, wrestler.id)
     if not stats:
         return
 
-    # Count matches in last 7 days
     try:
         week_ago = (datetime.strptime(game_date, "%Y-%m-%d") - timedelta(days=7)).strftime("%Y-%m-%d")
     except (ValueError, TypeError):
@@ -697,13 +508,13 @@ def update_conditioning(db: Session, wrestler: GameWrestlerDB, game_date: str):
         ).count()
     )
 
-    cond = stats.conditioning_level or 70
-    if matches_this_week >= 3:
-        cond = max(20, cond - 5)
+    cond = stats.conditioning_level or CONDITIONING_DEFAULT
+    if matches_this_week >= CONDITIONING_OVERWORK_THRESHOLD:
+        cond = max(CONDITIONING_MIN, cond - CONDITIONING_OVERWORK_PENALTY)
     elif matches_this_week == 0 and not wrestler.is_injured:
-        cond = min(100, cond + 3)
+        cond = min(MAX_STAT_CAP, cond + CONDITIONING_REST_GAIN)
     else:
-        cond = min(100, cond + 1)
+        cond = min(MAX_STAT_CAP, cond + CONDITIONING_WORK_GAIN)
     stats.conditioning_level = cond
 
 
@@ -711,20 +522,10 @@ def grow_specialization(stats: WrestlerStatsDB, stipulation: str):
     """Increase specialization from working a stipulation match."""
     if not stipulation:
         return
-
-    mapping = {
-        "cage": "cage_specialist",
-        "hell_in_a_cell": "cage_specialist",
-        "ladder": "ladder_specialist",
-        "tables": "hardcore_specialist",
-        "no_dq": "hardcore_specialist",
-        "extreme_rules": "hardcore_specialist",
-        "street_fight": "hardcore_specialist",
-    }
-    attr = mapping.get(stipulation.lower().replace(" ", "_"))
+    attr = STIPULATION_SPECIALIST_MAP.get(stipulation.lower().replace(" ", "_"))
     if attr:
         old = getattr(stats, attr, 0) or 0
-        setattr(stats, attr, min(100, old + 2))
+        setattr(stats, attr, min(MAX_STAT_CAP, old + SPECIALIZATION_GROWTH))
 
 
 # ---------------------------------------------------------------------------
@@ -744,7 +545,6 @@ def tick_persona(db: Session, world_id: str, game_date: str):
     ).all()
 
     for wrestler in wrestlers:
-        # Ensure persona data exists (migration for pre-existing wrestlers)
         backstory = db.query(WrestlerBackstoryDB).filter(
             WrestlerBackstoryDB.wrestler_id == wrestler.id,
         ).first()
@@ -759,13 +559,9 @@ def tick_persona(db: Session, world_id: str, game_date: str):
             persona_service.generate_initial_gimmick(db, wrestler, game_date)
             continue
 
-        # Tick gimmick staleness
         persona_service.tick_gimmick_staleness(db, wrestler, game_date)
-
-        # Evolve gimmick depth/fan investment based on recent activity
         persona_service.evolve_gimmick(db, wrestler, game_date)
 
-        # Check for repackaging pressure (NPC only)
         if wrestler.is_npc:
             pressure = persona_service.check_repackaging_pressure(db, wrestler)
             if pressure["pressure"] > 80:
@@ -773,10 +569,8 @@ def tick_persona(db: Session, world_id: str, game_date: str):
                     db, wrestler, game_date, pressure["reason"]
                 )
 
-        # Life event roll (~3% per wrestler per week)
         persona_service.generate_life_event(db, wrestler.id, world_id, game_date)
 
-        # Process effects of active life events
         active_events = db.query(LifeEventDB).filter(
             LifeEventDB.wrestler_id == wrestler.id,
             LifeEventDB.is_active == True,
